feat(install): add support for Homebrew on Linux with --with-brew option
This commit introduces an opt-in feature to install Homebrew on Linux by
adding the `--with-brew` flag. When this flag is specified, Homebrew will
be set up; otherwise, it is skipped on Linux systems.

Additionally, the help message has been updated to reflect the new option
and instructions on using Homebrew with the installation script.